pyproject.toml
[tool.poetry]
name = "dominantcolors"
version = "1.0.0"
description = "script to generate a palette of dominant colors from an image"
authors = ["pablonoya <pablonoya18@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
click = "^8.1.7"
numpy = "^2.0.1"
pillow-simd = "^9.5.0.post0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
dominantcolors = "dominantcolors.cli:cli"
